+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [0.13.0] - 2026-09-23
+
 ### Fixed
 
 - `--help` (`-h`) and `--version` now exit `0` with no